Update Launchpad_Data_Processor.ino
Made a few extra changes to the Xbee config:
- There is now an extra 'ATCN' command to exit command mode, meaning it starts transmitting data much more quickly
- I've added a timeout on the initial '+++' so that if the Xbee doesn't provide an OK the first time around, after 5 seconds, the '+++' is repeated

The only other issue is straight after power up, where the Xbee won't respond first time around. Shouldn't be a big deal to solve, but at the moment, a reset of the launchpad is still required to get it through the configuration after power up.