Update Launchpad_Data_Processor.ino
Added SD write routine on Matt's behalf, but it is now within another boolean condition so that we can still test without the SD connected.

Also cut a bunch of code that was commented out and updated the start/stop/reset string indicators. Starts in idle mode by default, so if using serial monitor type !!! to begin transmission.