Create CAN_Trans_Sim.ino
This simulates the format that the LV board should provide data in. I think its pretty much right, but I haven't been able to confirm that all data is correct on the receiver end. 

The bit shifting is a pain to work out, but its taken from the packet description we got from Daniel.