Create CAN_Receiver.ino
Pretty basic receiver just to provide a start point for integrating it with the main launchpad code.

Where I'm at at the moment is that I can switch on the transmitter then receiver, and the transmitter confirms data is no longer pending. However, I haven't been able to check the data on the receiver yet and I'm not sure whether that might be due to ID masking, or other problems in the configuration.